Implemented the Server-Database communication.
Here the Server gets the "configuration" request from the client
and queries the Database directly whether it has seen this
particular "performance" values and corresponding "configuration"
before. If the Database has seen this "configuration" then it will
respond back, further the "Server" responds back to the "Client"
with the final configuration.