Store code-server output in PID-based directory.
Every time the Active Harmony application starts the server, it
generates the PID(Process ID) for this particular run. Using the
PID, a folder with the name "harmony_PID" will be created under
"scratch" file repository. Now the newly generated code that the
code-servers are generating will be stored under the particular
PID folder.