Introduce atomic prefetch feature.
The PRO algorithm cannot fetch an arbitrary amount of candidate points
from the future.  It is strictly limited to the size of the search
simplex until the client returns a performance value for each
candidate point of the simplex.

An atomic queue only prefetches more points after receives N reports,
where N is the size of the queue.