Fixed two bugs in the Nelder-Mead algorithm.
BUGFIX1
Symptom: The NM algorithm does not seem to be converged forever.
Cause: When a simplex shrinked towards the best point, we should perform shrinking once and measure the performance at all the new points. However, the code estimated only one point rather than all the shrinked points, and the shrink step was incorrectly repeated forever. The nm_next_vertex function incorrectly set the index_curr variable to be 0 whenever the function executed a routine for the shrink step. The index_curr variable means which of the shrinked points should be tested next. Since index_curr could never reach the termination condition, the shrink step never finished.
Solution: I made the index_curr variable maintain a correct number. Also, I set index_curr to be -1 so that the nm_next_vertex function can notice the beginning of the shrink state. Then nm_next_vertex increments index_curr as one shrinked point is tested. So now all the shrinked points are tested, and the shrink step is not repeated forever. Eventually, the NM algorithm converges correctly.

BUGFIX2
Symptom: At a point when a simplex should be converged, an error occurred on a session server, and a harmony client produced the error message of "Failed to report performance to server."
Cause: There was a bug of missing "break;" at a switch-case statement in the nm_next_vertex function. So the nm_next_vertext function returns -1 to invoke an error even though the correct return value should be 0.
Solution: I simply added "break;" to the problem point. Now the error message disappeared.