Intoduce new documentation system.
Adds a new top-level makefile target called "doc".  This will use
doxygen to generate the User's Guide, Plug-in Writer's Guide, and
Internal Developer's Guide.  Each document is created in both PDF
and HTML formats.

This commit also introduces text for the User's Guide, and a
preliminary skeleton for the Plug-in Writer's Guide and Internal
Developer's Guide.