Add argc and argv parameters to harmony_init().
If provided to harmony_init(), argc and argv will be used look for
any Harmony configuration directives supplied on the command line.
If found, they will be removed and saved elsewhere to be applied
just prior to launch or after join, whichever comes first.  This
commit also bundles the following changes:

- Interpret escape characters when parsing configuration values.
- Force configuration keys to consist only of [0-9A-z_] characters.