Remove "static" from session-core global symbols.
The "--export-dynamic" linker flag is necessary to export functions
like session_getcfg() and session_error() to plug-ins.  However, the
flag is overzealous and adds all symbols to the dynamic symbol table.
The "static" modifier was used to reduce the number of symbols
included unnecessarily.  However, the resulting reduction (150 versus
196) does not outweigh the extra code.

We can consider using linker scripts if the dynamic symbol table
becomes problematically large.