Make layer init() functions re-invocable.
This represents a first pass at updating the existing layer plugins
to handle a session_restart().  A key difference between layer and
strategy init() functions is that layers can expect its fini()
function to be called prior to the re-invocation of its init()
function.

We should consider refactoring init() functions into a load()/unload()
pair for memory allocation, and a re-invokable init() for value
initialization only.