Add the HMESG_BEST message type.
Prior to this commit, the best known point was only sent to a client
along with the response to a fetch request.  This becomes a problem if
the strategy converges before any points are fetched, which is
possible when using the cache layer.

This commit introduces the HMESG_BEST message type that explicitly
requests the session's best known point.  This eliminates the need to
piggyback best point data during a fetch request.