Introduce the httpinfo plugin layer.
Reports session configuration information needed by the HTTP server.
This is the first step in eliminating a known HTTP server race
condition.

Currently, the HTTP server retrieves session information directly by
sending and receiving messages with a session process.  However, there
is no guarantee that the received message is a response to the
original message sent.